Back end/requirements.txt
chromadb==0.5.23
Flask==2.3.2
Flask_RESTful==0.3.9
nltk==3.9.1
numpy==2.2.0
Pillow==11.0.0
protobuf==5.29.2
pytz==2024.2
Werkzeug==3.1.3
google-generativeai==0.8.3
open-clip-torch==2.29.0